Backend/requirements.txt
Flask==2.2.2
flask-cors==3.1.1
cohere==4.1.1
numpy==1.24.1
pdfminer.six==20221105
pdfplumber==0.7.4
pytesseract==0.3.9
Pillow==9.4.0
pdf2image==1.16.3
faiss-cpu==1.7.3
PyPDF2==1.26.0
